WIP: Adds an untested assembleStorage for ELL
This is a somewhat dirty commit because it does more than one thing. In addition to providing an implementation of assembleStorage that compiles, it also changes some of the data types of the private attributes. Rather than using raw pointer arrays, I try to use the unique_pointers from the STL. The colIdx is also changed to an int from size_t.